refactor: 레포지토리 테스트에 테스트 격리를 위한 DatabaseCleaner 추가
